Reassign node output types and delete the variable contract layer
Names describing where a value came from become the shape they always
were: file paths and file content are strings, process and status
codes and header maps are integers or objects. Outputs and config
fields previously declared number are split into integer and float
according to whether the value can be fractional, and hotkey becomes
keyboard_key everywhere it still appeared.

The old contract table mapped several deleted names onto a handful of
loose buckets (numeric, text, file-path); with only the ten surviving
types left, matching a variable to a field is exact equality, so the
lookup table is gone in favor of a direct comparison. Config fields
that referenced the old buckets now declare their real type, and a
numeric field's accepted variable type follows its own declared kind
instead of a shared wildcard.

+++ data/nodes/definitions/actions/beep.ts
@@ -13,7 +13,7 @@ export const beepNode = defineNode({
 			label: "Frequency Hz",
 			type: "number",
 			usesVariables: true,
-			variableTypes: "numeric",
+			variableTypes: "float",
 			numeric: {
 				kind: "float",
 				signed: false,
@@ -28,7 +28,7 @@ export const beepNode = defineNode({
 			label: "Duration ms",
 			type: "number",
 			usesVariables: true,
-			variableTypes: "numeric",
+			variableTypes: "float",
 			numeric: {
 				kind: "float",
 				signed: false,
